+++ stringcompare/distance/__init__.py
@@ -1,4 +1,23 @@
-from stringcompare.distance._distance import Comparator, StringComparator, Levenshtein, DamerauLevenshtein, Jaro, JaroWinkler, LCSDistance
+from stringcompare.distance._distance import (
+    Comparator,
+    StringComparator,
+    Levenshtein,
+    DamerauLevenshtein,
+    Jaro,
+    JaroWinkler,
+    LCSDistance,
+    CharacterDifference,
+    Hamming,
+)
 
-__all__ = ["Comparator", "StringComparator", "Levenshtein",
-           "DamerauLevenshtein", "Jaro", "JaroWinkler", "LCSDistance"]
+__all__ = [
+    "Comparator",
+    "StringComparator",
+    "Levenshtein",
+    "DamerauLevenshtein",
+    "Jaro",
+    "JaroWinkler",
+    "LCSDistance",
+    "CharacterDifference",
+    "Hamming",
+]