+++ stringcompare/distance/comparator.py
@@ -1,19 +1,21 @@
 from abc import ABC, abstractmethod
 import numpy as np
+from typing import Generic, TypeVar, List
+
+T = TypeVar("T")
 
 
-class Comparator(ABC):
-
+class Comparator(ABC, Generic[T]):
     @abstractmethod
-    def compare(e1, e2):
+    def compare(s: T, t: T) -> float:
         """
         Comparison between two elements.
 
         Parameters
         ----------
-        e1:
+        s:
             element to compare from.
-        e1:
+        t:
             element to compare to.
 
         Returns
@@ -22,10 +24,10 @@ class Comparator(ABC):
         """
         pass
 
-    def __call__(self, e1, e2):
-        return self.compare(e1, e2)
+    def __call__(self, s: T, t: T) -> float:
+        return self.compare(s, t)
 
-    def pairwise(self, l1, l2):
+    def pairwise(self, l1: List[T], l2: List[T]) -> np.ndarray:
         """
         Pairwise comparisons between two lists.
 
@@ -38,17 +40,18 @@ class Comparator(ABC):
 
         Returns
         -------
-        Matrix of dimension len(l1)xlen(l2), where each row corresponds to an element of l1 and each column corresponds to an element of l2.
+        Matrix of dimension len(l1)xlen(l2), where each row corresponds to an
+        element of l1 and each column corresponds to an element of l2.
         """
         return np.array([[self.compare(s, t, self.dmat) for t in l2] for s in l1])
 
-    def elementwise(self, l1, l2):
+    def elementwise(self, l1: List[T], l2: List[T]) -> np.ndarray:
         return np.array([self.compare(s, t, self.dmat) for s, t in zip(l1, l2)])
 
 
-class StringComparator(Comparator):
+class StringComparator(Comparator[str]):
     pass
 
 
-class NumericComparator(Comparator):
+class NumericComparator(Comparator[float]):
     pass