+++ stringcompare/distance/lcs.hpp
@@ -9,9 +9,6 @@
 namespace py = pybind11;
 using namespace std;
 
-template<class T>
-using Mat = vector<vector<T>>;
-
 class LCSDistance: public StringComparator {
 public:
 
@@ -19,7 +16,7 @@ public:
   bool similarity;
   int dmat_size;
   bool check_bounds;
-  Mat<int> dmat;
+  vector<int> dmat;
 
   LCSDistance(bool normalize=true, bool similarity=false, int dmat_size=100, bool check_bounds=true){
     this->normalize = normalize;
@@ -27,7 +24,7 @@ public:
     this->dmat_size = dmat_size;
     this->check_bounds = check_bounds;
 
-    dmat = Mat<int>(2, vector<int>(dmat_size));
+    dmat = vector<int>(dmat_size);
   }
 
   int lcs(const string &s, const string &t) {
@@ -35,30 +32,31 @@ public:
     int n = t.size();
 
     for (int i = 0; i < dmat_size; i++) {
-      dmat[0][i] = 0;
+      dmat[i] = 0;
     }
 
+    int p = 0;
+    int temp;
     for (int j = 1; j <= n; j++) {
-      dmat[(j-1) % 2][0] = 0;
-      dmat[j % 2][1] = 0;
+      temp = 0;
+      p = 0;
       for (int i = 1; i <= m; i++) {
         if (s[i-1] != t[j-1]){
-          dmat[j % 2][i] = std::max({dmat[(j-1) % 2][i], dmat[j % 2][i-1]});
+          p = std::max({dmat[i], p});
         } else {
-            dmat[j % 2][i] = dmat[(j-1) % 2][i-1] + 1;
+          p = temp + 1;
         }
-        
+        temp = dmat[i];
+        dmat[i] = p;
       }
     }
 
-    return dmat[n % 2][m];
+    return p;
   }
 
   double compare(const string &s, const string &t) {
     if (check_bounds) {
-      auto m = max(s.size(), t.size()) + 1;
-      dmat[0].reserve(m);
-      dmat[1].reserve(m);
+      dmat.reserve(s.size()+1);
     }
 
     double len = s.size() + t.size();