+++ CHANGELOG.rst
@@ -1,11 +1,15 @@
-=======
+=========
 Changelog
-=======
+=========
 
 Unreleased
 ----------
 
-
+* Update contribution guidelines and development workflow
+* Refactor and update documentation and development files
+* Update pybind11 package documentation
+* Improve memory efficiency of Levenshtein and LCSDistance classes
+* Add CharacterDifference class
 
 0.0.1 (2022-01-18)
 ------------------